+++ src/www/ble-com-protocol.ts
@@ -8,12 +8,15 @@
 import { QueueComProtocol } from '@iotize/device-client.js/protocol/impl/queue-com-protocol';
 import { ComProtocolConnectOptions, ComProtocolDisconnectOptions, ComProtocolSendOptions } from '@iotize/device-client.js/protocol/api/com-protocol.interface';
 import { FormatHelper } from '@iotize/device-client.js/core/format/format-helper';
-import { from, Observable } from 'rxjs';
+import { from, Observable, Subscriber, Subscription, Subject } from 'rxjs';
+import { first } from "rxjs/operators";
 import { CordovaInterface } from './cordova-interface';
 
 declare var iotizeBLE: CordovaInterface;
 
 export class BLEComProtocol extends QueueComProtocol {
+    private _connectionStateSubject?: Subject<any>;
+    _connectionChangeObservable: any;
 
     log(level: 'info' | 'error' | 'debug', ...args: any[]): void {
         console[level](level.toUpperCase() + " [IoTizeBLEProtocol] | " + args.map(entry => {
@@ -26,6 +29,9 @@ export class BLEComProtocol extends QueueComProtocol {
         }).join(" "));
     }
 
+    private _connectionStateSubscription?: Subscription;
+    private _connectionObservable?: Observable<any>;
+
    private deviceId: string = "";
 
    constructor(deviceId: string) {
@@ -37,10 +43,30 @@ export class BLEComProtocol extends QueueComProtocol {
 
     _connect(options?: ComProtocolConnectOptions): Observable<any> {
         this.log('info', '_connect', options);
-        return from(this._cordovaCallToPromise(
-            iotizeBLE.connect,
-            this.deviceId
-        ));
+        // return from(this._cordovaCallToPromise(
+        //     iotizeBLE.connect,
+        //     this.deviceId
+        // ));
+
+        if (!this._connectionStateSubject) {
+            this._connectionStateSubject = new Subject();
+        }
+
+        this._connectionStateSubscription = this._connectionStateSubject.subscribe((val) => this.setConnectionState(val));
+
+        const onConnect = (val: any) => {
+            this.log('debug', '_connect observable: onConnect');
+            this._connectionStateSubject.next(val);
+        };
+        const onError = (error: any) => {
+            this.log('debug', '_connect observable: onError');
+            this._connectionStateSubject.error(error);
+        };
+        iotizeBLE.connect(this.deviceId, onConnect ,onError);
+
+        return this._connectionStateSubject.pipe(
+            first()
+        );
     }
 
 
@@ -87,6 +113,22 @@ export class BLEComProtocol extends QueueComProtocol {
             cordovaFct.apply(iotizeBLE, args);
         });
     }
+
+    // protected _connectionChangeObservable(): Observable<any>{
+    //     this.log('info', '_connectionChangeObservable');
+
+    //     return Observable.create((subscriber: Subscriber<any>) => {
+    //         const onConnectionChange = (val: any) => {
+    //             this.log('debug', '_connect observable: onConnect');
+    //             subscriber.next(val);
+    //         };
+    //         const onError = (error: any) => {
+    //             this.log('debug', '_connect observable: onError');
+    //             subscriber.error(error);
+    //         };
+    //         iotizeBLE.connectionChangeListener(onConnectionChange ,onError);
+    //     });
+    // }
 
  };