+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.1.0-alpha.23"></a>
+- fix: iOS : CheckAvailable now uses actual bluetooth manager state
+
 <a name="1.1.0-alpha.21"></a>
 # [1.1.0-alpha.21](https://bitbucket.org/iotize-solutions/device-com-ble.cordova/compare/v1.1.0-alpha.20...v1.1.0-alpha.21) (2019-04-04)
 - fix: Scan no longer needs active location to work